// lib/api/generated/models/models-create-case-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UtilsNotificationEntityType } from './utils-notification-entity-type';
// May contain unused imports in some cases
// @ts-ignore
import type { UtilsPriority } from './utils-priority';

export interface ModelsCreateCaseRequest {
    'description': string;
    'priority'?: UtilsPriority;
    'related_entity_id'?: string;
    'related_entity_type'?: UtilsNotificationEntityType;
    'title': string;
}